+++ lib/plausible_web/controllers/api/stats_controller.ex
@@ -695,37 +695,29 @@ defmodule PlausibleWeb.Api.StatsController do
   end
 
   def referrer_drilldown(conn, %{"referrer" => "Google"} = params) do
-    site = conn.assigns[:site] |> Repo.preload(:google_auth)
+    site = conn.assigns[:site]
 
-    query =
-      Query.from(site, params)
-      |> Query.put_filter([:is, "visit:source", "Google"])
-
-    # :TODO: Goal exclusion doesn't work here!
-    search_terms =
-      if site.google_auth && site.google_auth.property && !Query.get_filter(query, "goal") do
-        google_api().fetch_stats(site, query, params["limit"] || 9)
-      end
-
-    %{:visitors => %{value: total_visitors}} = Stats.aggregate(site, query, [:visitors])
+    query = Query.from(site, params)
 
     user_id = get_session(conn, :current_user_id)
     is_admin = user_id && Plausible.Sites.has_admin_access?(user_id, site)
 
-    case search_terms do
-      nil ->
-        json(conn, %{not_configured: true, is_admin: is_admin, total_visitors: total_visitors})
+    case google_api().fetch_stats(site, query, params["limit"] || 9) do
+      {:error, :google_propery_not_configured} ->
+        json(conn, %{not_configured: true, is_admin: is_admin})
+
+      {:error, :unsupported_filters} ->
+        json(conn, %{unsupported_filters: true})
 
       {:ok, terms} ->
-        json(conn, %{search_terms: terms, total_visitors: total_visitors})
+        json(conn, %{search_terms: terms})
 
       {:error, _} ->
         conn
         |> put_status(502)
         |> json(%{
           not_configured: true,
-          is_admin: is_admin,
-          total_visitors: total_visitors
+          is_admin: is_admin
         })
     end
   end