+++ test/plausible/google/api_test.exs
@@ -245,7 +245,7 @@ defmodule Plausible.Google.APITest do
           "https://www.googleapis.com/webmasters/v3/sites/sc-domain%3Adummy.test/searchAnalytics/query",
           [{"Authorization", "Bearer 123"}],
           %{
-            dimensionFilterGroups: %{},
+            dimensionFilterGroups: [],
             dimensions: ["query"],
             endDate: "2022-01-05",
             rowLimit: 5,
@@ -301,82 +301,96 @@ defmodule Plausible.Google.APITest do
     end
   end
 
-  describe "fetch_stats/3 with VCR cassetes" do
-    test "returns name and visitor count", %{user: user, site: site} do
+  test "returns error when token refresh fails", %{user: user, site: site} do
+    mock_http_with("google_analytics_auth#invalid_grant.json")
+
+    insert(:google_auth,
+      user: user,
+      site: site,
+      property: "sc-domain:dummy.test",
+      access_token: "*****",
+      refresh_token: "*****",
+      expires: NaiveDateTime.add(NaiveDateTime.utc_now(), -3600)
+    )
+
+    query = %Plausible.Stats.Query{date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])}
+
+    assert {:error, "invalid_grant"} = Google.API.fetch_stats(site, query, 5)
+  end
+
+  test "returns error when google auth not configured", %{site: site} do
+    query = %Plausible.Stats.Query{date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])}
+
+    assert {:error, :google_property_not_configured} = Google.API.fetch_stats(site, query, 5)
+  end
+
+  describe "fetch_stats/3 with valid auth" do
+    setup %{user: user, site: site} do
+      insert(:google_auth,
+        user: user,
+        site: site,
+        property: "sc-domain:dummy.test",
+        expires: NaiveDateTime.add(NaiveDateTime.utc_now(), 3600)
+      )
+
+      :ok
+    end
+
+    test "returns name and visitor count", %{site: site} do
       mock_http_with("google_analytics_stats.json")
 
-      insert(:google_auth,
-        user: user,
-        site: site,
-        property: "sc-domain:dummy.test",
-        expires: NaiveDateTime.add(NaiveDateTime.utc_now(), 3600)
-      )
-
       query = %Plausible.Stats.Query{date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])}
 
       assert {:ok,
               [
-                %{name: ["keyword1", "keyword2"], visitors: 25},
-                %{name: ["keyword3", "keyword4"], visitors: 15}
+                %{name: "keyword1", visitors: 25, ctr: 36.8, impressions: 50, position: 2.2},
+                %{name: "keyword3", visitors: 15}
               ]} = Google.API.fetch_stats(site, query, 5)
     end
 
-    test "returns next page when page argument is set", %{user: user, site: site} do
-      mock_http_with("google_analytics_stats#with_page.json")
-
-      insert(:google_auth,
-        user: user,
-        site: site,
-        property: "sc-domain:dummy.test",
-        expires: NaiveDateTime.add(NaiveDateTime.utc_now(), 3600)
+    test "transforms page filters to search console format", %{site: site} do
+      expect(
+        Plausible.HTTPClient.Mock,
+        :post,
+        fn
+          "https://www.googleapis.com/webmasters/v3/sites/sc-domain%3Adummy.test/searchAnalytics/query",
+          [{"Authorization", "Bearer 123"}],
+          %{
+            dimensionFilterGroups: [
+              %{filters: [%{expression: "https://dummy.test/page", dimension: "page"}]}
+            ],
+            dimensions: ["query"],
+            endDate: "2022-01-05",
+            rowLimit: 5,
+            startDate: "2022-01-01"
+          } ->
+            {:ok, %Finch.Response{status: 200, body: %{"rows" => []}}}
+        end
       )
 
-      query = %Plausible.Stats.Query{
-        filters: [[:is, "page", 5]],
-        date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])
-      }
+      query =
+        Plausible.Stats.Query.from(site, %{
+          "period" => "custom",
+          "from" => "2022-01-01",
+          "to" => "2022-01-05",
+          "filters" => "event:page==/page"
+        })
 
-      assert {:ok,
-              [
-                %{name: ["keyword1", "keyword2"], visitors: 25},
-                %{name: ["keyword3", "keyword4"], visitors: 15}
-              ]} = Google.API.fetch_stats(site, query, 5)
+      assert {:ok, []} = Google.API.fetch_stats(site, query, 5)
     end
 
-    test "defaults first page when page argument is not set", %{user: user, site: site} do
-      mock_http_with("google_analytics_stats#without_page.json")
+    test "returns :invalid filters when using filters that cannot be used in Search Console", %{
+      site: site
+    } do
+      query =
+        Plausible.Stats.Query.from(site, %{
+          "period" => "custom",
+          "from" => "2022-01-01",
+          "to" => "2022-01-05",
+          "filters" => "event:goal==Signup"
+        })
 
-      insert(:google_auth,
-        user: user,
-        site: site,
-        property: "sc-domain:dummy.test",
-        expires: NaiveDateTime.add(NaiveDateTime.utc_now(), 3600)
-      )
-
-      query = %Plausible.Stats.Query{date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])}
-
-      assert {:ok,
-              [
-                %{name: ["keyword1", "keyword2"], visitors: 25},
-                %{name: ["keyword3", "keyword4"], visitors: 15}
-              ]} = Google.API.fetch_stats(site, query, 5)
-    end
-
-    test "returns error when token refresh fails", %{user: user, site: site} do
-      mock_http_with("google_analytics_auth#invalid_grant.json")
-
-      insert(:google_auth,
-        user: user,
-        site: site,
-        property: "sc-domain:dummy.test",
-        access_token: "*****",
-        refresh_token: "*****",
-        expires: NaiveDateTime.add(NaiveDateTime.utc_now(), -3600)
-      )
-
-      query = %Plausible.Stats.Query{date_range: Date.range(~D[2022-01-01], ~D[2022-01-05])}
-
-      assert {:error, "invalid_grant"} = Google.API.fetch_stats(site, query, 5)
+      assert {:error, :unsupported_filters} = Google.API.fetch_stats(site, query, 5)
     end
   end